Heavily reworked command line options and reporting.
Command line options reduced down to 3 and the coverage config file
can be used to set other behaviour.

Everyone wants combined reports even in dist=each mode, so multiple
reports are removed and it always combines into a single report.  This
also nicely simplifies the code.

Uses the latest in development coverage which added support for the
source arg used to specify the roots of source trees to measure.