Remove the automatic SIGTERM handler install from the afterfork handler and update docs.
The reason for this change being that multiprocessing
is going to be broken on windows or pypy anyway, so this is best left to
the users to deal in their own code. The documentation includes a
working pattern to use (close/join). If the users really needs to use
terminate then they should have a platform check and compromise between
lower coverage measurement on pypy or chance to deadlock (as seen in the
travis failures in https://github.com/pytest-dev/pytest-cov/pull/265).